requirements.txt

langchain
langchain-community
langchain-text-splitters
langchain-huggingface
chromadb
sentence-transformers
neo4j
python-dotenv
pymupdf
streamlit
huggingface_hub
transformers
accelerate
spacy
 